+++ docs/release_notes.rst
@@ -10,7 +10,7 @@ Release notes
 * Fixed issues with mouse gaze data collection (multiple gaze points when no RBS button clicked)
 * Refactored code for analyzers and data loaders.
 * Added full set of readings from eyetracker (not limited to average pupil size, or average gaze point)
-* Added HeatMapAnalyzer complete code
+* Added HeatMapAnalyzer, FixationAnalyzer, FocusMapAnalyzer, SaccadeAnalyzer complete code
 
 
 0.1.1 (2025-10-28)